Add test for slide samples to dataset
