+++ tests/metadata/dicom_schema/sample/test_dicom_sample.py
@@ -12,61 +12,35 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-import datetime
-from typing import List, Optional, Sequence, Union
+from typing import List, Optional, Sequence
 from pydicom import Dataset
 import pytest
 
 from pydicom.uid import UID
 from pydicom.sr.coding import Code
 from tests.metadata.dicom_schema.helpers import (
-    assert_item_measurement_equals_value,
-    assert_item_name_equals_code,
-    assert_item_string_equals_value,
     assert_next_item_equals_code,
-    assert_next_item_equals_datetime,
     assert_next_item_equals_measurement,
     assert_next_item_equals_string,
-    create_collection_dataset,
     create_description_dataset,
-    create_measurement_item,
-    create_processing_dataset,
-    create_receiving_dataset,
     create_sampling_dataset,
-    create_staining_dataset,
-    create_storage_dataset,
-    create_string_item,
+    create_specimen_preparation_sequence,
 )
 from wsidicom.conceptcode import (
     ContainerTypeCode,
     SpecimenCollectionProcedureCode,
     SpecimenEmbeddingMediaCode,
     SpecimenFixativesCode,
-    SpecimenPreparationStepsCode,
     SpecimenStainsCode,
     SpecimenSamplingProcedureCode,
     AnatomicPathologySpecimenTypesCode,
-    UnitCode,
 )
-from wsidicom.metadata.schema.dicom.sample.model import (
-    CollectionDicomModel,
-    ProcessingDicomModel,
-    StainingDicomModel,
-    SamplingDicomModel,
-    ReceivingDicomModel,
-    StorageDicomModel,
-)
+from wsidicom.metadata.schema.dicom.sample.formatter import SpecimenDicomFormatter
+from wsidicom.metadata.schema.dicom.sample.model import SamplingDicomModel
 from wsidicom.metadata.schema.dicom.sample.parser import SpecimenDicomParser
 from wsidicom.metadata.schema.dicom.sample.schema import (
-    CollectionDicomSchema,
-    ProcessingDicomSchema,
-    ReceivingDicomSchema,
-    SamplingDicomSchema,
-    SpecimenDescriptionDicomSchema,
-    SpecimenLocalizationDicomSchema,
-    StainingDicomSchema,
     SampleCodes,
-    StorageDicomSchema,
+    SpecimenDescriptionDicomSchema,
 )
 
 from wsidicom.metadata.sample import (
@@ -74,290 +48,13 @@ from wsidicom.metadata.sample import (
     Embedding,
     Specimen,
     Fixation,
-    LocalIssuerOfIdentifier,
-    Measurement,
     Sample,
     SamplingLocation,
     SlideSample,
     SpecimenLocalization,
-    SpecimenIdentifier,
     Sampling,
-    Processing,
     Staining,
-    Receiving,
-    Storage,
 )
-
-
-@pytest.fixture()
-def specimen_id():
-    yield "specimen"
-
-
-@pytest.fixture()
-def slide_sample_id():
-    yield "slide sample"
-
-
-@pytest.fixture()
-def slide_sample_uid():
-    yield UID("1.2.826.0.1.3680043.8.498.11522107373528810886192809691753445424")
-
-
-@pytest.fixture()
-def specimen_type():
-    yield AnatomicPathologySpecimenTypesCode("tissue specimen")
-
-
-@pytest.fixture()
-def collection_method():
-    yield SpecimenCollectionProcedureCode("Excision")
-
-
-@pytest.fixture()
-def specimen_sampling_method():
-    yield SpecimenSamplingProcedureCode("Dissection")
-
-
-@pytest.fixture()
-def block_id():
-    yield "block"
-
-
-@pytest.fixture()
-def block_sampling_method():
-    yield SpecimenSamplingProcedureCode("Block sectioning")
-
-
-@pytest.fixture()
-def block_type():
-    yield AnatomicPathologySpecimenTypesCode("tissue specimen")
-
-
-@pytest.fixture()
-def specimen_localization():
-    yield SpecimenLocalization(description="left")
-
-
-@pytest.fixture()
-def sampling_location():
-    yield SamplingLocation(description="left")
-
-
-@pytest.fixture()
-def primary_anatomic_structures():
-    yield [Code("value", "schema", "meaning")]
-
-
-@pytest.fixture()
-def container():
-    yield ContainerTypeCode("Specimen container")
-
-
-@pytest.fixture()
-def stains():
-    yield [
-        SpecimenStainsCode("hematoxylin stain"),
-        SpecimenStainsCode("water soluble eosin stain"),
-    ]
-
-
-@pytest.fixture()
-def collection_dicom(
-    collection: Collection,
-    identifier: Union[str, SpecimenIdentifier],
-    fixative: Optional[SpecimenFixativesCode],
-    medium: Optional[SpecimenEmbeddingMediaCode],
-    processing_method: Optional[SpecimenPreparationStepsCode],
-    container: Optional[ContainerTypeCode],
-    specimen_type: Optional[AnatomicPathologySpecimenTypesCode],
-):
-    identifier, issuer = SpecimenIdentifier.get_string_identifier_and_issuer(identifier)
-    yield CollectionDicomModel(
-        identifier=identifier,
-        issuer_of_identifier=issuer,
-        date_time=collection.date_time,
-        description=collection.description,
-        fixative=fixative,
-        embedding=medium,
-        method=collection.method,
-        processing=processing_method,
-        container=container,
-        specimen_type=specimen_type,
-    )
-
-
-@pytest.fixture()
-def collection_dataset(
-    collection_dicom: CollectionDicomModel,
-    identifier: Union[str, SpecimenIdentifier],
-):
-    yield create_collection_dataset(collection_dicom, identifier)
-
-
-@pytest.fixture()
-def sampling_dicom(
-    sampling: Sampling,
-    identifier: Union[str, SpecimenIdentifier],
-    fixative: Optional[SpecimenFixativesCode],
-    medium: Optional[SpecimenEmbeddingMediaCode],
-    processing_method: Optional[SpecimenPreparationStepsCode],
-    container: Optional[ContainerTypeCode],
-    specimen_type: Optional[AnatomicPathologySpecimenTypesCode],
-):
-    identifier, issuer = SpecimenIdentifier.get_string_identifier_and_issuer(identifier)
-    (
-        parent_identifier,
-        parent_issuer,
-    ) = SpecimenIdentifier.get_string_identifier_and_issuer(
-        sampling.specimen.identifier
-    )
-    yield SamplingDicomModel(
-        identifier=identifier,
-        issuer_of_identifier=issuer,
-        date_time=sampling.date_time,
-        description=sampling.description,
-        method=sampling.method,
-        fixative=fixative,
-        embedding=medium,
-        parent_specimen_identifier=parent_identifier,
-        issuer_of_parent_specimen_identifier=parent_issuer,
-        parent_specimen_type=sampling.specimen_type,
-        processing=processing_method,
-        location_reference=sampling.location.reference
-        if sampling.location is not None
-        else None,
-        location_description=sampling.location.description
-        if sampling.location is not None
-        else None,
-        location_x=sampling.location.x if sampling.location is not None else None,
-        location_y=sampling.location.y if sampling.location is not None else None,
-        location_z=sampling.location.z if sampling.location is not None else None,
-        container=container,
-        specimen_type=specimen_type,
-    )
-
-
-@pytest.fixture()
-def sampling_dataset(
-    sampling_dicom: SamplingDicomModel,
-    identifier: Union[str, SpecimenIdentifier],
-):
-    yield create_sampling_dataset(sampling_dicom, identifier)
-
-
-@pytest.fixture()
-def processing_dicom(
-    processing: Processing,
-    identifier: Union[str, SpecimenIdentifier],
-    fixative: Optional[SpecimenFixativesCode],
-    medium: Optional[SpecimenEmbeddingMediaCode],
-):
-    identifier, issuer = SpecimenIdentifier.get_string_identifier_and_issuer(identifier)
-    yield ProcessingDicomModel(
-        identifier=identifier,
-        issuer_of_identifier=issuer,
-        date_time=processing.date_time,
-        description=processing.description,
-        processing=processing.method,
-        fixative=fixative,
-        embedding=medium,
-        container=None,
-        specimen_type=None,
-    )
-
-
-@pytest.fixture()
-def processing_dataset(
-    processing_dicom: ProcessingDicomModel,
-    identifier: Union[str, SpecimenIdentifier],
-):
-    yield create_processing_dataset(processing_dicom, identifier)
-
-
-@pytest.fixture()
-def staining_dicom(
-    staining: Staining,
-    identifier: Union[str, SpecimenIdentifier],
-    fixative: Optional[SpecimenFixativesCode],
-    medium: Optional[SpecimenEmbeddingMediaCode],
-    processing_method: Optional[SpecimenPreparationStepsCode],
-):
-    identifier, issuer = SpecimenIdentifier.get_string_identifier_and_issuer(identifier)
-    yield StainingDicomModel(
-        identifier=identifier,
-        issuer_of_identifier=issuer,
-        date_time=staining.date_time,
-        description=staining.description,
-        substances=staining.substances,
-        fixative=fixative,
-        embedding=medium,
-        processing=processing_method,
-        container=None,
-        specimen_type=None,
-    )
-
-
-@pytest.fixture()
-def staining_dataset(
-    staining_dicom: StainingDicomModel,
-    identifier: Union[str, SpecimenIdentifier],
-):
-    yield create_staining_dataset(staining_dicom, identifier)
-
-
-@pytest.fixture()
-def receiving_dicom(
-    receiving: Receiving,
-    identifier: Union[str, SpecimenIdentifier],
-):
-    identifier, issuer = SpecimenIdentifier.get_string_identifier_and_issuer(identifier)
-    yield ReceivingDicomModel(
-        identifier=identifier,
-        issuer_of_identifier=issuer,
-        date_time=receiving.date_time,
-        description=receiving.description,
-        fixative=None,
-        embedding=None,
-        processing=None,
-        container=None,
-        specimen_type=None,
-    )
-
-
-@pytest.fixture()
-def receiving_dataset(
-    receiving_dicom: ReceivingDicomModel,
-    identifier: Union[str, SpecimenIdentifier],
-):
-    yield create_receiving_dataset(receiving_dicom, identifier)
-
-
-@pytest.fixture()
-def storage_dicom(
-    storage: Storage,
-    identifier: Union[str, SpecimenIdentifier],
-):
-    identifier, issuer = SpecimenIdentifier.get_string_identifier_and_issuer(identifier)
-    yield StorageDicomModel(
-        identifier=identifier,
-        issuer_of_identifier=issuer,
-        date_time=storage.date_time,
-        description=storage.description,
-        fixative=None,
-        embedding=None,
-        processing=None,
-        container=None,
-        specimen_type=None,
-    )
-
-
-@pytest.fixture()
-def storage_dataset(
-    storage_dicom: StorageDicomModel,
-    identifier: Union[str, SpecimenIdentifier],
-):
-    yield create_storage_dataset(storage_dicom, identifier)
 
 
 class TestSampleDicom:
@@ -427,9 +124,8 @@ class TestSampleDicom:
         for slide_sample_id, slide_sample_uid, specimen_id in zip(
             slide_sample_ids, slide_sample_uids, specimen_ids
         ):
-            description = create_description_dataset(
+            preparation_steps = create_specimen_preparation_sequence(
                 slide_sample_id,
-                slide_sample_uid,
                 specimen_id,
                 block_id,
                 specimen_type,
@@ -439,14 +135,19 @@ class TestSampleDicom:
                 medium,
                 block_sampling_method,
                 block_type,
-                specimen_localization,
                 sampling_location,
-                primary_anatomic_structures,
                 stains,
+                specimen_container,
+                block_container,
+            )
+            description = create_description_dataset(
+                slide_sample_id,
+                slide_sample_uid,
+                primary_anatomic_structures,
+                specimen_localization,
                 short_description=short_description,
                 detailed_description=detailed_description,
-                specimen_container=specimen_container,
-                block_container=block_container,
+                preparation_step_datasets=preparation_steps,
             )
 
             descriptions.append(description)
@@ -498,941 +199,385 @@ class TestSampleDicom:
                 assert isinstance(collection_step, Collection)
                 assert collection_step.method == collection_method
 
-
-class TestPreparationStepDicomSchema:
     @pytest.mark.parametrize(
-        ["date_time", "description"],
+        ["slide_sample_ids", "slide_sample_uids", "specimen_ids"],
         [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    @pytest.mark.parametrize(
-        "container", [None, ContainerTypeCode("Specimen container")]
-    )
-    @pytest.mark.parametrize(
-        "specimen_type", [None, AnatomicPathologySpecimenTypesCode("tissue specimen")]
-    )
-    def test_serialize_collection_dicom(
-        self,
-        collection_dicom: CollectionDicomModel,
-        identifier: Union[str, SpecimenIdentifier],
-    ):
-        # Arrange
-        schema = CollectionDicomSchema()
-
-        # Act
-        serialized = schema.dump(collection_dicom)
-
-        # Assert
-        assert isinstance(serialized, list)
-        item_iterator = iter(serialized)
-        # First item should be identifier
-        assert_next_item_equals_string(
-            item_iterator,
-            SampleCodes.identifier,
-            identifier if isinstance(identifier, str) else identifier.value,
-        )
-        # Next item can be issuer of identifier
-        if isinstance(identifier, SpecimenIdentifier) and identifier.issuer is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.issuer_of_identifier,
-                identifier.issuer.to_hl7v2(),
-            )
-        # Next item can be container type
-        if collection_dicom.container is not None:
-            assert_next_item_equals_code(
-                item_iterator, SampleCodes.container, collection_dicom.container
-            )
-        # Next item can be specimen type
-        if collection_dicom.specimen_type is not None:
-            assert_next_item_equals_code(
-                item_iterator,
-                SampleCodes.specimen_type,
-                collection_dicom.specimen_type,
-            )
-        # Next item should be processing type
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.processing_type,
-            SampleCodes.specimen_collection,
-        )
-        # Next item can be date time
-        if collection_dicom.date_time is not None:
-            assert_next_item_equals_datetime(
-                item_iterator,
-                SampleCodes.datetime_of_processing,
-                collection_dicom.date_time,
-            )
-        # Next item can be description
-        if collection_dicom.description is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.processing_description,
-                collection_dicom.description,
-            )
-        # Last item should be method
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.specimen_collection,
-            collection_dicom.method,
-        )
-        # There should be no more items
-        assert next(item_iterator, None) is None
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    @pytest.mark.parametrize(
-        ["fixative", "medium"],
-        [
-            [None, None],
             [
-                SpecimenFixativesCode("Neutral Buffered Formalin"),
-                SpecimenEmbeddingMediaCode("Paraffin wax"),
+                ["slide sample"],
+                [
+                    UID(
+                        "1.2.826.0.1.3680043.8.498.11522107373528810886192809691753445424"
+                    )
+                ],
+                ["specimen"],
+            ],
+            [
+                ["slide sample 1", "slide sample 2"],
+                [
+                    UID(
+                        "1.2.826.0.1.3680043.8.498.11522107373528810886192809691753445424"
+                    ),
+                    UID(
+                        "1.2.826.0.1.3680043.8.498.11522107373528810886192809691753445425"
+                    ),
+                ],
+                ["specimen 1", "specimen 2"],
             ],
         ],
     )
     @pytest.mark.parametrize(
-        "container", [None, ContainerTypeCode("Specimen container")]
+        "fixative", [SpecimenFixativesCode("Neutral Buffered Formalin")]
+    )
+    @pytest.mark.parametrize("medium", [SpecimenEmbeddingMediaCode("Paraffin wax")])
+    @pytest.mark.parametrize("short_description", [None, "short description"])
+    @pytest.mark.parametrize("detailed_description", [None, "detailed description"])
+    @pytest.mark.parametrize(
+        "specimen_container", [None, ContainerTypeCode("Specimen container")]
     )
     @pytest.mark.parametrize(
-        "specimen_type", [None, AnatomicPathologySpecimenTypesCode("tissue specimen")]
+        "block_container", [None, ContainerTypeCode("Tissue cassette")]
     )
-    def test_deserialize_collection_dicom(
+    def test_slide_sample_to_dataset(
         self,
-        collection_dicom: CollectionDicomModel,
-        collection_dataset: Dataset,
+        slide_sample_ids: Sequence[str],
+        slide_sample_uids: Sequence[UID],
+        specimen_ids: Sequence[str],
+        block_id: str,
+        specimen_type: AnatomicPathologySpecimenTypesCode,
+        collection_method: SpecimenCollectionProcedureCode,
+        fixative: SpecimenFixativesCode,
+        specimen_sampling_method: SpecimenSamplingProcedureCode,
+        medium: SpecimenEmbeddingMediaCode,
+        block_sampling_method: SpecimenSamplingProcedureCode,
+        block_type: AnatomicPathologySpecimenTypesCode,
+        specimen_localization: SpecimenLocalization,
+        sampling_location: SamplingLocation,
+        primary_anatomic_structures: Sequence[Code],
+        stains: Sequence[SpecimenStainsCode],
+        short_description: Optional[str],
+        detailed_description: Optional[str],
+        specimen_container: Optional[ContainerTypeCode],
+        block_container: Optional[ContainerTypeCode],
     ):
         # Arrange
-        schema = CollectionDicomSchema()
+        slide_samples = []
+        for slide_sample_id, slide_sample_uid, specimen_id in zip(
+            slide_sample_ids, slide_sample_uids, specimen_ids
+        ):
+            specimen = Specimen(
+                identifier=specimen_id,
+                extraction_step=Collection(method=collection_method),
+                type=specimen_type,
+                container=specimen_container,
+                steps=[
+                    Fixation(fixative=fixative),
+                ],
+            )
+            block = Sample(
+                identifier=block_id,
+                sampled_from=[
+                    specimen.sample(
+                        method=specimen_sampling_method,
+                        location=sampling_location,
+                    )
+                ],
+                type=block_type,
+                container=block_container,
+                steps=[
+                    Embedding(medium=medium),
+                ],
+            )
+            slide_sample = SlideSample(
+                identifier=slide_sample_id,
+                uid=slide_sample_uid,
+                anatomical_sites=primary_anatomic_structures,
+                localization=specimen_localization,
+                sampled_from=block.sample(
+                    method=block_sampling_method,
+                ),
+                short_description=short_description,
+                detailed_description=detailed_description,
+            )
+            slide_samples.append(slide_sample)
+        staining = Staining(stains)
+
+        schema = SpecimenDescriptionDicomSchema()
 
         # Act
-        deserialized = schema.load(
-            collection_dataset.SpecimenPreparationStepContentItemSequence
-        )
+        descriptions = []
+        for slide_sample in slide_samples:
+            dicom_model = SpecimenDicomFormatter.to_dicom(slide_sample, [staining])
+            description = schema.dump(dicom_model)
+            descriptions.append(description)
+
+        dataset = Dataset()
+        dataset.SpecimenDescriptionSequence = descriptions
 
         # Assert
-        assert isinstance(deserialized, CollectionDicomModel)
-        assert deserialized.identifier == collection_dicom.identifier
-        assert (
-            deserialized.issuer_of_identifier == collection_dicom.issuer_of_identifier
-        )
-        assert deserialized.method == collection_dicom.method
-        assert deserialized.date_time == collection_dicom.date_time
-        assert deserialized.description == collection_dicom.description
-        assert deserialized.fixative == collection_dicom.fixative
-        assert deserialized.embedding == collection_dicom.embedding
-        assert deserialized.container == collection_dicom.container
-        assert deserialized.specimen_type == collection_dicom.specimen_type
+        for index, (slide_sample_id, slide_sample_uid, specimen_id) in enumerate(
+            zip(slide_sample_ids, slide_sample_uids, specimen_ids)
+        ):
+            description = dataset.SpecimenDescriptionSequence[index]
+            assert isinstance(description, Dataset)
+            assert description.SpecimenIdentifier == slide_sample_id
+            assert description.SpecimenUID == slide_sample_uid
+            assert description.SpecimenShortDescription == short_description
+            assert description.SpecimenDetailedDescription == detailed_description
+            assert len(description.PrimaryAnatomicStructureSequence) == len(
+                primary_anatomic_structures
+            )
+            for index, primary_anatomic_structure in enumerate(
+                primary_anatomic_structures
+            ):
+                assert (
+                    description.PrimaryAnatomicStructureSequence[index].CodeValue
+                    == primary_anatomic_structure.value
+                )
+                assert (
+                    description.PrimaryAnatomicStructureSequence[
+                        index
+                    ].CodingSchemeDesignator
+                    == primary_anatomic_structure.scheme_designator
+                )
+                assert (
+                    description.PrimaryAnatomicStructureSequence[index].CodeMeaning
+                    == primary_anatomic_structure.meaning
+                )
 
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    @pytest.mark.parametrize(
-        "location",
-        [
-            None,
-            SpecimenLocalization(
-                "reference",
-                "description",
-                Measurement(1, UnitCode("cm")),
-                Measurement(2, UnitCode("cm")),
-                Measurement(3, UnitCode("cm")),
-            ),
-        ],
-    )
-    @pytest.mark.parametrize(
-        "container", [None, ContainerTypeCode("Specimen container")]
-    )
-    @pytest.mark.parametrize(
-        "specimen_type", [None, AnatomicPathologySpecimenTypesCode("tissue specimen")]
-    )
-    def test_serialize_sampling_dicom(
-        self,
-        sampling_dicom: SamplingDicomModel,
-        identifier: Union[str, SpecimenIdentifier],
-    ):
-        # Arrange
-        schema = SamplingDicomSchema()
+            step_iterator = iter(description.SpecimenPreparationSequence)
 
-        # Act
-        serialized = schema.dump(sampling_dicom)
-
-        # Assert
-        assert isinstance(serialized, list)
-        item_iterator = iter(serialized)
-        # First item should be identifier
-        assert_next_item_equals_string(
-            item_iterator,
-            SampleCodes.identifier,
-            identifier if isinstance(identifier, str) else identifier.value,
-        )
-        # Next item can be issuer of identifier
-        if isinstance(identifier, SpecimenIdentifier) and identifier.issuer is not None:
+            # Collection
+            step = next(step_iterator)
+            step_item_iterator = iter(step.SpecimenPreparationStepContentItemSequence)
             assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.issuer_of_identifier,
-                identifier.issuer.to_hl7v2(),
+                step_item_iterator, SampleCodes.identifier, specimen_id
             )
-        # Next item can be container type
-        if sampling_dicom.container is not None:
-            assert_next_item_equals_code(
-                item_iterator, SampleCodes.container, sampling_dicom.container
-            )
-        # Next item can be specimen type
-        if sampling_dicom.specimen_type is not None:
-            assert_next_item_equals_code(
-                item_iterator,
-                SampleCodes.specimen_type,
-                sampling_dicom.specimen_type,
-            )
-        # Next item should be processing type
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.processing_type,
-            SampleCodes.sampling_method,
-        )
-        # Next item can be date time
-        if sampling_dicom.date_time is not None:
-            assert_next_item_equals_datetime(
-                item_iterator,
-                SampleCodes.datetime_of_processing,
-                sampling_dicom.date_time,
-            )
-        # Next item can be description
-        if sampling_dicom.description is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.processing_description,
-                sampling_dicom.description,
-            )
-        # Next item should be method
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.sampling_method,
-            sampling_dicom.method,
-        )
-        # Next item should be parent specimen identifier
-        assert_next_item_equals_string(
-            item_iterator,
-            SampleCodes.parent_specimen_identifier,
-            sampling_dicom.parent_specimen_identifier,
-        )
-        # Next item can be parent specimen identifier issuer
-        if sampling_dicom.issuer_of_parent_specimen_identifier is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.issuer_of_parent_specimen_identifier,
-                sampling_dicom.issuer_of_parent_specimen_identifier,
-            )
-        # Next item should be parent specimen type
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.parent_specimen_type,
-            sampling_dicom.parent_specimen_type,
-        )
-        # Next item can be location reference
-        if sampling_dicom.location_reference is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.location_frame_of_reference,
-                sampling_dicom.location_reference,
-            )
-        # Next item can be location description
-        if sampling_dicom.location_description is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.location_of_sampling_site,
-                sampling_dicom.location_description,
-            )
-        # Next item can be location x
-        if sampling_dicom.location_x is not None:
-            assert_next_item_equals_measurement(
-                item_iterator,
-                SampleCodes.location_of_sampling_site_x,
-                sampling_dicom.location_x,
-            )
-        # Next item can be location y
-        if sampling_dicom.location_y is not None:
-            assert_next_item_equals_measurement(
-                item_iterator,
-                SampleCodes.location_of_sampling_site_y,
-                sampling_dicom.location_y,
-            )
-        # Next item can be location z
-        if sampling_dicom.location_z is not None:
-            assert_next_item_equals_measurement(
-                item_iterator,
-                SampleCodes.location_of_sampling_site_z,
-                sampling_dicom.location_z,
-            )
-        # There should be no more items
-        assert next(item_iterator, None) is None
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    @pytest.mark.parametrize(
-        ["fixative", "medium"],
-        [
-            [None, None],
-            [
-                SpecimenFixativesCode("Neutral Buffered Formalin"),
-                SpecimenEmbeddingMediaCode("Paraffin wax"),
-            ],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "location",
-        [
-            None,
-            SpecimenLocalization(
-                "reference",
-                "description",
-                Measurement(1, UnitCode("cm")),
-                Measurement(2, UnitCode("cm")),
-                Measurement(3, UnitCode("cm")),
-            ),
-        ],
-    )
-    @pytest.mark.parametrize(
-        "container", [None, ContainerTypeCode("Specimen container")]
-    )
-    @pytest.mark.parametrize(
-        "specimen_type", [None, AnatomicPathologySpecimenTypesCode("tissue specimen")]
-    )
-    def test_deserialize_sampling_dicom(
-        self, sampling_dicom: SamplingDicomModel, sampling_dataset: Dataset
-    ):
-        # Arrange
-        schema = SamplingDicomSchema()
-
-        # Act
-        deserialized = schema.load(
-            sampling_dataset.SpecimenPreparationStepContentItemSequence
-        )
-
-        # Assert
-        assert isinstance(deserialized, SamplingDicomModel)
-        assert deserialized.identifier == sampling_dicom.identifier
-        assert deserialized.issuer_of_identifier == sampling_dicom.issuer_of_identifier
-        assert deserialized.method == sampling_dicom.method
-        assert deserialized.date_time == sampling_dicom.date_time
-        assert deserialized.description == sampling_dicom.description
-        assert deserialized.fixative == sampling_dicom.fixative
-        assert deserialized.embedding == sampling_dicom.embedding
-        assert deserialized.location_reference == sampling_dicom.location_reference
-        assert deserialized.location_description == sampling_dicom.location_description
-        assert deserialized.location_x == sampling_dicom.location_x
-        assert deserialized.location_y == sampling_dicom.location_y
-        assert deserialized.location_z == sampling_dicom.location_z
-        assert deserialized.container == sampling_dicom.container
-        assert deserialized.specimen_type == sampling_dicom.specimen_type
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    @pytest.mark.parametrize(
-        ["fixative", "medium"],
-        [
-            [None, None],
-            [
-                SpecimenFixativesCode("Neutral Buffered Formalin"),
-                SpecimenEmbeddingMediaCode("Paraffin wax"),
-            ],
-        ],
-    )
-    def test_serialize_processing_dicom(
-        self,
-        processing_dicom: ProcessingDicomModel,
-        identifier: Union[str, SpecimenIdentifier],
-    ):
-        # Arrange
-        schema = ProcessingDicomSchema()
-
-        # Act
-        serialized = schema.dump(processing_dicom)
-
-        # Assert
-        assert isinstance(serialized, list)
-        item_iterator = iter(serialized)
-        # First item should be identifier
-        assert_next_item_equals_string(
-            item_iterator,
-            SampleCodes.identifier,
-            identifier if isinstance(identifier, str) else identifier.value,
-        )
-        # Next item can be issuer of identifier
-        if isinstance(identifier, SpecimenIdentifier) and identifier.issuer is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.issuer_of_identifier,
-                identifier.issuer.to_hl7v2(),
-            )
-        # Next item should be processing type
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.processing_type,
-            SampleCodes.sample_processing,
-        )
-        # Next item can be date time
-        if processing_dicom.date_time is not None:
-            assert_next_item_equals_datetime(
-                item_iterator,
-                SampleCodes.datetime_of_processing,
-                processing_dicom.date_time,
-            )
-        # Next item can be description
-        if processing_dicom.description is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.processing_description,
-                processing_dicom.description,
-            )
-        # Next item can be method
-        if processing_dicom.processing is not None:
-            assert_next_item_equals_code(
-                item_iterator,
-                SampleCodes.processing_description,
-                processing_dicom.processing,
-            )
-        # Next item can be fixative
-        if processing_dicom.fixative is not None:
-            assert_next_item_equals_code(
-                item_iterator, SampleCodes.fixative, processing_dicom.fixative
-            )
-        # Next item can be embedding
-        if processing_dicom.embedding is not None:
-            assert_next_item_equals_code(
-                item_iterator, SampleCodes.embedding, processing_dicom.embedding
-            )
-        # There should be no more items
-        assert next(item_iterator, None) is None
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    @pytest.mark.parametrize(
-        ["fixative", "medium"],
-        [
-            [None, None],
-            [
-                SpecimenFixativesCode("Neutral Buffered Formalin"),
-                SpecimenEmbeddingMediaCode("Paraffin wax"),
-            ],
-        ],
-    )
-    def test_deserialize_processing_dicom(
-        self,
-        processing_dicom: ProcessingDicomModel,
-        processing_dataset: Dataset,
-    ):
-        # Arrange
-
-        schema = ProcessingDicomSchema()
-
-        # Act
-        deserialized = schema.load(
-            processing_dataset.SpecimenPreparationStepContentItemSequence
-        )
-
-        # Assert
-        assert isinstance(deserialized, ProcessingDicomModel)
-        assert deserialized.identifier == processing_dicom.identifier
-        assert (
-            deserialized.issuer_of_identifier == processing_dicom.issuer_of_identifier
-        )
-        assert deserialized.processing == processing_dicom.processing
-        assert deserialized.date_time == processing_dicom.date_time
-        assert deserialized.description == processing_dicom.description
-        assert deserialized.fixative == processing_dicom.fixative
-        assert deserialized.embedding == processing_dicom.embedding
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    def test_serialize_staining_dicom(
-        self,
-        staining_dicom: StainingDicomModel,
-        identifier: Union[str, SpecimenIdentifier],
-    ):
-        # Arrange
-        schema = StainingDicomSchema()
-
-        # Act
-        serialized = schema.dump(staining_dicom)
-
-        # Assert
-        assert isinstance(serialized, list)
-        item_iterator = iter(serialized)
-        item_iterator = iter(serialized)
-        # First item should be identifier
-        assert_next_item_equals_string(
-            item_iterator,
-            SampleCodes.identifier,
-            identifier if isinstance(identifier, str) else identifier.value,
-        )
-        # Next item can be issuer of identifier
-        if isinstance(identifier, SpecimenIdentifier) and identifier.issuer is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.issuer_of_identifier,
-                identifier.issuer.to_hl7v2(),
-            )
-        # Next item should be processing type
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.processing_type,
-            SampleCodes.staining,
-        )
-        # Next item can be date time
-        if staining_dicom.date_time is not None:
-            assert_next_item_equals_datetime(
-                item_iterator,
-                SampleCodes.datetime_of_processing,
-                staining_dicom.date_time,
-            )
-        # Next item can be description
-        if staining_dicom.description is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.processing_description,
-                staining_dicom.description,
-            )
-        # Next item can be staining
-        for substance in staining_dicom.substances:
-            if isinstance(substance, SpecimenStainsCode):
+            if specimen_container is not None:
                 assert_next_item_equals_code(
-                    item_iterator, SampleCodes.using_substance, substance
+                    step_item_iterator, SampleCodes.container, specimen_container
                 )
-            else:
-                assert_next_item_equals_string(
-                    item_iterator, SampleCodes.using_substance, substance
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.specimen_type, specimen_type
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.processing_type,
+                SampleCodes.specimen_collection,
+            )
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.specimen_collection, collection_method
+            )
+            assert next(step_item_iterator, None) is None
+
+            # Fixation
+            step = next(step_iterator)
+            step_item_iterator = iter(step.SpecimenPreparationStepContentItemSequence)
+            assert_next_item_equals_string(
+                step_item_iterator, SampleCodes.identifier, specimen_id
+            )
+            if specimen_container is not None:
+                assert_next_item_equals_code(
+                    step_item_iterator, SampleCodes.container, specimen_container
                 )
-        # There should be no more items
-        assert next(item_iterator, None) is None
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.specimen_type, specimen_type
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.processing_type,
+                SampleCodes.sample_processing,
+            )
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.fixative, fixative
+            )
+            assert next(step_item_iterator, None) is None
 
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    def test_deserialize_staining_dicom(
-        self,
-        staining_dicom: StainingDicomModel,
-        staining_dataset: Dataset,
-    ):
-        # Arrange
-
-        schema = StainingDicomSchema()
-
-        # Act
-        deserialized = schema.load(
-            staining_dataset.SpecimenPreparationStepContentItemSequence
-        )
-
-        # Assert
-        assert isinstance(deserialized, StainingDicomModel)
-        assert deserialized.identifier == staining_dicom.identifier
-        assert deserialized.issuer_of_identifier == staining_dicom.issuer_of_identifier
-        assert deserialized.substances == staining_dicom.substances
-        assert deserialized.date_time == staining_dicom.date_time
-        assert deserialized.description == staining_dicom.description
-        assert deserialized.fixative == staining_dicom.fixative
-        assert deserialized.embedding == staining_dicom.embedding
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    def test_serialize_receiving_dicom(
-        self,
-        receiving_dicom: ReceivingDicomModel,
-        identifier: Union[str, SpecimenIdentifier],
-    ):
-        # Arrange
-        schema = ReceivingDicomSchema()
-
-        # Act
-        serialized = schema.dump(receiving_dicom)
-
-        # Assert
-        assert isinstance(serialized, list)
-        item_iterator = iter(serialized)
-        # First item should be identifier
-        assert_next_item_equals_string(
-            item_iterator,
-            SampleCodes.identifier,
-            identifier if isinstance(identifier, str) else identifier.value,
-        )
-        # Next item can be issuer of identifier
-        if isinstance(identifier, SpecimenIdentifier) and identifier.issuer is not None:
+            # Sampling to block
+            step = next(step_iterator)
+            step_item_iterator = iter(step.SpecimenPreparationStepContentItemSequence)
             assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.issuer_of_identifier,
-                identifier.issuer.to_hl7v2(),
+                step_item_iterator, SampleCodes.identifier, block_id
             )
-        # Next item should be processing type
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.processing_type,
-            SampleCodes.receiving,
-        )
-        # Next item can be date time
-        if receiving_dicom.date_time is not None:
-            assert_next_item_equals_datetime(
-                item_iterator,
-                SampleCodes.datetime_of_processing,
-                receiving_dicom.date_time,
-            )
-        # Next item can be description
-        if receiving_dicom.description is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.processing_description,
-                receiving_dicom.description,
-            )
-        # There should be no more items
-        assert next(item_iterator, None) is None
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    def test_deserialize_receiving_dicom(
-        self,
-        receiving_dicom: ReceivingDicomModel,
-        receiving_dataset: Dataset,
-    ):
-        # Arrange
-
-        schema = ReceivingDicomSchema()
-
-        # Act
-        deserialized = schema.load(
-            receiving_dataset.SpecimenPreparationStepContentItemSequence
-        )
-
-        # Assert
-        assert isinstance(deserialized, ReceivingDicomModel)
-        assert deserialized.identifier == receiving_dicom.identifier
-        assert deserialized.issuer_of_identifier == receiving_dicom.issuer_of_identifier
-        assert deserialized.date_time == receiving_dicom.date_time
-        assert deserialized.description == receiving_dicom.description
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    def test_serialize_storage_dicom(
-        self,
-        storage_dicom: StorageDicomModel,
-        identifier: Union[str, SpecimenIdentifier],
-    ):
-        # Arrange
-        schema = StorageDicomSchema()
-
-        # Act
-        serialized = schema.dump(storage_dicom)
-
-        # Assert
-        assert isinstance(serialized, list)
-        item_iterator = iter(serialized)
-        # First item should be identifier
-        assert_next_item_equals_string(
-            item_iterator,
-            SampleCodes.identifier,
-            identifier if isinstance(identifier, str) else identifier.value,
-        )
-        # Next item can be issuer of identifier
-        if isinstance(identifier, SpecimenIdentifier) and identifier.issuer is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.issuer_of_identifier,
-                identifier.issuer.to_hl7v2(),
-            )
-        # Next item should be processing type
-        assert_next_item_equals_code(
-            item_iterator,
-            SampleCodes.processing_type,
-            SampleCodes.storage,
-        )
-        # Next item can be date time
-        if storage_dicom.date_time is not None:
-            assert_next_item_equals_datetime(
-                item_iterator,
-                SampleCodes.datetime_of_processing,
-                storage_dicom.date_time,
-            )
-        # Next item can be description
-        if storage_dicom.description is not None:
-            assert_next_item_equals_string(
-                item_iterator,
-                SampleCodes.processing_description,
-                storage_dicom.description,
-            )
-        # There should be no more items
-        assert next(item_iterator, None) is None
-
-    @pytest.mark.parametrize(
-        ["date_time", "description"],
-        [
-            [datetime.datetime(2021, 1, 1), "description"],
-            [None, None],
-        ],
-    )
-    @pytest.mark.parametrize(
-        "identifier",
-        [
-            "identifier",
-            SpecimenIdentifier("identifier", LocalIssuerOfIdentifier("issuer")),
-        ],
-    )
-    def test_deserialize_storage_dicom(
-        self,
-        storage_dicom: StorageDicomModel,
-        storage_dataset: Dataset,
-    ):
-        # Arrange
-
-        schema = StorageDicomSchema()
-
-        # Act
-        deserialized = schema.load(
-            storage_dataset.SpecimenPreparationStepContentItemSequence
-        )
-
-        # Assert
-        assert isinstance(deserialized, StorageDicomModel)
-        assert deserialized.identifier == storage_dicom.identifier
-        assert deserialized.issuer_of_identifier == storage_dicom.issuer_of_identifier
-        assert deserialized.date_time == storage_dicom.date_time
-        assert deserialized.description == storage_dicom.description
-
-    @pytest.mark.parametrize("reference", ["reference", None])
-    @pytest.mark.parametrize("description", ["description", None])
-    @pytest.mark.parametrize("x", [Measurement(1, UnitCode("mm")), None])
-    @pytest.mark.parametrize("y", [Measurement(1, UnitCode("mm")), None])
-    @pytest.mark.parametrize("z", [Measurement(1, UnitCode("mm")), None])
-    @pytest.mark.parametrize("visual_marking", ["visual_marking", None])
-    def test_serialize_specimen_location_dicom(
-        self,
-        reference: Optional[str],
-        description: Optional[str],
-        x: Optional[Measurement],
-        y: Optional[Measurement],
-        z: Optional[Measurement],
-        visual_marking: Optional[str],
-    ):
-        # Arrange
-        location = SpecimenLocalization(
-            reference=reference,
-            description=description,
-            x=x,
-            y=y,
-            z=z,
-            visual_marking=visual_marking,
-        )
-        schema = SpecimenLocalizationDicomSchema()
-
-        # Act
-        serialized = schema.dump(location)
-
-        # Assert
-        assert isinstance(serialized, list)
-        item_iterator = iter(serialized)
-        for string, name in [
-            (reference, SampleCodes.location_frame_of_reference),
-            (description, SampleCodes.location_of_specimen),
-        ]:
-            if string is not None:
-                string_item = next(item_iterator)
-                assert_item_name_equals_code(string_item, name)
-                assert_item_string_equals_value(string_item, string)
-        for measurement, name in [
-            (x, SampleCodes.location_of_specimen_x),
-            (y, SampleCodes.location_of_specimen_y),
-            (z, SampleCodes.location_of_specimen_z),
-        ]:
-            if measurement is not None:
-                measurement_item = next(item_iterator)
-                assert_item_name_equals_code(measurement_item, name)
-                assert_item_measurement_equals_value(measurement_item, measurement)
-
-        if visual_marking is not None:
-            visual_marking_item = next(item_iterator)
-            assert_item_name_equals_code(
-                visual_marking_item, SampleCodes.visual_marking_of_specimen
-            )
-            assert_item_string_equals_value(visual_marking_item, visual_marking)
-
-    @pytest.mark.parametrize("reference", ["reference", None])
-    @pytest.mark.parametrize("description", ["description", None])
-    @pytest.mark.parametrize("x", [Measurement(1, UnitCode("mm")), None])
-    @pytest.mark.parametrize("y", [Measurement(1, UnitCode("mm")), None])
-    @pytest.mark.parametrize("z", [Measurement(1, UnitCode("mm")), None])
-    @pytest.mark.parametrize("visual_marking", ["visual_marking", None])
-    def test_deserialize_specimen_location_dicom(
-        self,
-        reference: Optional[str],
-        description: Optional[str],
-        x: Optional[Measurement],
-        y: Optional[Measurement],
-        z: Optional[Measurement],
-        visual_marking: Optional[str],
-    ):
-        # Arrange
-        sequence = []
-        if reference is not None:
-            sequence.append(
-                create_string_item(SampleCodes.location_frame_of_reference, reference)
-            )
-        if description is not None:
-            sequence.append(
-                create_string_item(SampleCodes.location_of_specimen, description)
-            )
-        if x is not None:
-            sequence.append(
-                create_measurement_item(SampleCodes.location_of_specimen_x, x)
-            )
-        if y is not None:
-            sequence.append(
-                create_measurement_item(SampleCodes.location_of_specimen_y, y)
-            )
-        if z is not None:
-            sequence.append(
-                create_measurement_item(SampleCodes.location_of_specimen_z, z)
-            )
-        if visual_marking is not None:
-            sequence.append(
-                create_string_item(
-                    SampleCodes.visual_marking_of_specimen, visual_marking
+            if block_container is not None:
+                assert_next_item_equals_code(
+                    step_item_iterator, SampleCodes.container, block_container
                 )
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.specimen_type, block_type
             )
-        schema = SpecimenLocalizationDicomSchema()
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.processing_type,
+                SampleCodes.sampling_method,
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.sampling_method,
+                specimen_sampling_method,
+            )
+            assert_next_item_equals_string(
+                step_item_iterator, SampleCodes.parent_specimen_identifier, specimen_id
+            )
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.parent_specimen_type, specimen_type
+            )
+            if sampling_location is not None:
+                if sampling_location.reference is not None:
+                    assert_next_item_equals_string(
+                        step_item_iterator,
+                        SampleCodes.location_frame_of_reference,
+                        sampling_location.reference,
+                    )
+                if sampling_location.description is not None:
+                    assert_next_item_equals_string(
+                        step_item_iterator,
+                        SampleCodes.location_of_sampling_site,
+                        sampling_location.description,
+                    )
+                if sampling_location.x is not None:
+                    assert_next_item_equals_measurement(
+                        step_item_iterator,
+                        SampleCodes.location_of_sampling_site_x,
+                        sampling_location.x,
+                    )
+                if sampling_location.y is not None:
+                    assert_next_item_equals_measurement(
+                        step_item_iterator,
+                        SampleCodes.location_of_sampling_site_y,
+                        sampling_location.y,
+                    )
+                if sampling_location.z is not None:
+                    assert_next_item_equals_measurement(
+                        step_item_iterator,
+                        SampleCodes.location_of_sampling_site_z,
+                        sampling_location.z,
+                    )
+            assert next(step_item_iterator, None) is None
+
+            # Embedding
+            step = next(step_iterator)
+            step_item_iterator = iter(step.SpecimenPreparationStepContentItemSequence)
+            assert_next_item_equals_string(
+                step_item_iterator, SampleCodes.identifier, block_id
+            )
+            if block_container is not None:
+                assert_next_item_equals_code(
+                    step_item_iterator, SampleCodes.container, block_container
+                )
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.specimen_type, block_type
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.processing_type,
+                SampleCodes.sample_processing,
+            )
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.embedding, medium
+            )
+            assert next(step_item_iterator, None) is None
+
+            # Sampling to slide sample
+            step = next(step_iterator)
+            step_item_iterator = iter(step.SpecimenPreparationStepContentItemSequence)
+            assert_next_item_equals_string(
+                step_item_iterator, SampleCodes.identifier, slide_sample_id
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.container,
+                ContainerTypeCode("Microscope slide"),
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.specimen_type,
+                AnatomicPathologySpecimenTypesCode("Slide"),
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.processing_type,
+                SampleCodes.sampling_method,
+            )
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.sampling_method, block_sampling_method
+            )
+            assert_next_item_equals_string(
+                step_item_iterator, SampleCodes.parent_specimen_identifier, block_id
+            )
+            assert_next_item_equals_code(
+                step_item_iterator, SampleCodes.parent_specimen_type, block_type
+            )
+            assert next(step_item_iterator, None) is None
+
+            # Staining
+            step = next(step_iterator)
+            step_item_iterator = iter(step.SpecimenPreparationStepContentItemSequence)
+            assert_next_item_equals_string(
+                step_item_iterator, SampleCodes.identifier, slide_sample_id
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.container,
+                ContainerTypeCode("Microscope slide"),
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.specimen_type,
+                AnatomicPathologySpecimenTypesCode("Slide"),
+            )
+            assert_next_item_equals_code(
+                step_item_iterator,
+                SampleCodes.processing_type,
+                SampleCodes.staining,
+            )
+            for stain in stains:
+                assert_next_item_equals_code(
+                    step_item_iterator, SampleCodes.using_substance, stain
+                )
+            assert next(step_item_iterator, None) is None
+
+    def test_sampling_from_not_defined_specimen(
+        self,
+        slide_sample_id: str,
+        slide_sample_uid: UID,
+        block_id: str,
+        block_sampling_method: SpecimenSamplingProcedureCode,
+        block_type: AnatomicPathologySpecimenTypesCode,
+    ):
+        # Arrange
+        sampling = create_sampling_dataset(
+            SamplingDicomModel(
+                identifier=slide_sample_id,
+                method=block_sampling_method,
+                parent_specimen_identifier=block_id,
+                parent_specimen_type=block_type,
+            ),
+            identifier=slide_sample_id,
+        )
+        description = create_description_dataset(
+            slide_sample_id,
+            slide_sample_uid,
+            preparation_step_datasets=[sampling],
+        )
+        dataset = Dataset()
+        dataset.SpecimenDescriptionSequence = [description]
+
+        schema = SpecimenDescriptionDicomSchema()
 
         # Act
-        deserialized = schema.load(sequence)
+        models = [
+            schema.load(description)
+            for description in dataset.SpecimenDescriptionSequence
+        ]
+        slide_samples, stainings = SpecimenDicomParser().parse_descriptions(models)
 
         # Assert
-        assert isinstance(deserialized, SpecimenLocalization)
-        assert deserialized.reference == reference
-        assert deserialized.description == description
-        assert deserialized.x == x
-        assert deserialized.y == y
-        assert deserialized.z == z
-        assert deserialized.visual_marking == visual_marking
+        assert len(stainings) == 0
+        assert len(slide_samples) == 1
+        slide_sample = slide_samples[0]
+        assert slide_sample.sampled_from is not None
+        assert isinstance(slide_sample.sampled_from, Sampling)
+        assert slide_sample.sampled_from.method == block_sampling_method
+        assert isinstance(slide_sample.sampled_from.specimen, Specimen)
+        assert slide_sample.sampled_from.specimen.identifier == block_id
+        assert slide_sample.sampled_from.specimen.type == block_type
